{{ cookiecutter.repo_name }}/requirements.txt
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
keras
theano
pandas
numpy
scipy
h5py
matplotlib
